Modifier les import pour que ca fonctionne et add la bdd dans le backend avec les endpoints

+++ app/src/main/java/com/example/kev/projetkevloic/Database/ObserverDB.java
@@ -215,9 +215,9 @@ public class ObserverDB {
 
         for (com.example.kev.myapplication.backend.observationApi.model.Observation p : observs) {
             ContentValues values = new ContentValues();
-            values.put(dbHelper.OBSERVATION_ID, p.getId());
-            values.put(dbHelper.OBSERVATION_ORNITHO, p.getOrni());
-            values.put(dbHelper.OBSERVATION_OISEAU, p.getOiseau());
+            values.put(dbHelper.OBSERVATION_ID,(long) p.getId());
+            values.put(dbHelper.OBSERVATION_ORNITHO, (long) p.getOrni());
+            values.put(dbHelper.OBSERVATION_OISEAU, (long) p.getOiseau());
             values.put(dbHelper.OBSERVATION_TEXT, p.getText());
